refactor(StudentDetails): Move StudentDetailsDisplay into its own file
